Front end Auth updates
Detail page will update based on whether or not the user is logged in. began work on the index page for this.

Things to implement-

if user is logged in, they will not be able to view the login page or the sign up page.

If the user is NOT logged in they will not be able to view the account pages or the create pages